lots of changes including the creation of templates, relocation of html files, accessibility changes, major transition to flask, etc.
