--- gui/test.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>About PassFraze</title>
-    <link rel="stylesheet" href="style.css">
-</head>
-<body>
-    <header>
-        <div class="header_content">
-            <h1><a href="index.html">PassFraze</a></h1>
-            <nav>
-                <a href="about.html" class="header_buttons">About</a>
-                <a href="faq.html" class="header_buttons">FAQ</a>
-                <a href="index.html" class="header_button_3">4 in 1 Tool</a>
-            </nav>
-        </div>
-    </header>
