+++ templates/about.html
@@ -4,15 +4,15 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>About PassFraze</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="gui/style.css">
 </head>
 <body>
     <header>
         <div class="header_content">
-            <h1><a href="index.html">PassFraze</a></h1>
+            <h1><a href="/">PassFraze</a></h1>
             <nav>
-                <a href="faq.html" class="header_buttons">FAQ</a>
-                <a href="index.html" class="header_button_3">4 in 1 Tool</a>
+                <a href="/faq" class="header_buttons">FAQ</a>
+                <a href="/" class="header_button_3">4 in 1 Tool</a>
             </nav>
         </div>
     </header>
@@ -24,7 +24,7 @@
                 personal accounts or fortifying your business's digital assets, 
                 PassFraze is your reliable companion in the realm of passwords.</p>
             <h1 style="text-align: left; font-size: 1.45rem;">Four Essential Capabilities</h1>
-            <img src="four.svg">
+            <img src="gui/four.svg">
             <h1 style="text-align: left; font-size: 1.15rem;">1. Password Generation</h1>
             <p>Create robust, complex passwords effortlessly with PassFraze's 
                 advanced password generator. Tailor-made with your preferences for both personal
