+++ templates/crack.html
@@ -4,15 +4,16 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>About PassFraze</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="gui/style.css">
 </head>
 <body>
     <header>
         <div class="header_content">
-            <h1><a href="index.html">PassFraze</a></h1>
+            <h1><a href="/">PassFraze</a></h1>
             <nav>
-                <a href="about.html" class="header_buttons">About</a>
-                <a href="index.html" class="header_button_3">4 in 1 Tool</a>
+                <a href="/about" class="header_buttons">About</a>
+                <a href="/faq" class="header_buttons">FAQ</a>
+                <a href="/" class="header_button_3">4 in 1 Tool</a>
             </nav>
         </div>
     </header>
