+++ templates/faq.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>About PassFraze</title>
+    <link rel="stylesheet" href="gui/style.css">
+</head>
+<body>
+    <header>
+        <div class="header_content">
+            <h1><a href="/">PassFraze</a></h1>
+            <nav>
+                <a href="/about" class="header_buttons">About</a>
+                <a href="/" class="header_button_3">4 in 1 Tool</a>
+            </nav>
+        </div>
+    </header>
