+++ templates/generate.html
@@ -4,16 +4,16 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>About PassFraze</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="gui/style.css">
 </head>
 <body>
     <header>
         <div class="header_content">
-            <h1><a href="index.html">PassFraze</a></h1>
+            <h1><a href="/">PassFraze</a></h1>
             <nav>
-                <a href="about.html" class="header_buttons">About</a>
-                <a href="faq.html" class="header_buttons">FAQ</a>
-                <a href="index.html" class="header_button_3">4 in 1 Tool</a>
+                <a href="/about" class="header_buttons">About</a>
+                <a href="/faq" class="header_buttons">FAQ</a>
+                <a href="/" class="header_button_3">4 in 1 Tool</a>
             </nav>
         </div>
         
@@ -23,16 +23,16 @@
         <label for="length">Password Length:</label>
         <input type="number" id="length" name="length" min="1" value="12"><br>
 
-        <input type="checkbox" id="uppercase" name="uppercase" checked>
+        <input type="checkbox" id="uppercase" name="uppercase">
         <label for="uppercase">Include Uppercase</label><br>
 
-        <input type="checkbox" id="lowercase" name="lowercase" checked>
+        <input type="checkbox" id="lowercase" name="lowercase">
         <label for="lowercase">Include Lowercase</label><br>
 
-        <input type="checkbox" id="digits" name="digits" checked>
+        <input type="checkbox" id="digits" name="digits">
         <label for="digits">Include Digits</label><br>
 
-        <input type="checkbox" id="special" name="special" checked>
+        <input type="checkbox" id="special" name="special">
         <label for="special">Include Special Characters</label><br>
 
         <button type="submit">Generate Password</button>
