+++ templates/index.html
@@ -4,16 +4,16 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>PassFraze</title>
-    <link rel="stylesheet" href="style.css">
-    <script src="main.js" defer></script>
+    <link rel="stylesheet" href="gui/style.css">
+    <script src="../gui/main.js" defer></script>
 </head>
 <body>
     <header>
         <div class="header_content">
-            <h1><a href="index.html">PassFraze</a></h1>
+            <h1><a href="/">PassFraze</a></h1>
             <nav>
-                <a href="about.html" class="header_buttons">About</a>
-                <a href="faq.html" class="header_buttons">FAQ</a>
+                <a href="/about" class="header_buttons">About</a>
+                <a href="/faq" class="header_buttons">FAQ</a>
             </nav>
         </div>
     </header>
@@ -23,7 +23,6 @@
                 <button id="test">Test Password Strength</button>
                 <button id="crack">Crack Password</button>
             </section>
-            <script src="main.js"></script>
         </div>
     </main>
 </body>
